Marconi.CLI

Documentation

chainPointParser :: Parser ChainPoint

multiString :: Mod OptionFields [Address ShelleyAddr] -> Parser TargetAddresses

parses CLI params to valid NonEmpty list of Shelley addresses We error out if there are any invalid addresses

parseCardanoAddresses :: String -> [Address ShelleyAddr]

pNetworkId :: Parser NetworkId

data Options

This executable is meant to exercise a set of indexers (for now datumhash -> datum) against the mainnet (meant to be used for testing).

In case you want to access the results of the datumhash indexer you need to query the resulting database: $ sqlite3 datums.sqlite > select slotNo, datumHash, datum from kv_datumhsh_datum where slotNo = 39920450; 39920450|679a55b523ff8d61942b2583b76e5d49498468164802ef1ebe513c685d6fb5c2|X(002f9787436835852ea78d3c45fc3d436b324184

Constructors

Options

Fields

optionsSocketPath :: String

optionsNetworkId :: NetworkId

optionsChainPoint :: ChainPoint

optionsDbPath :: FilePath

SQLite database directory path

optionsDisableUtxo :: Bool

optionsDisableDatum :: Bool

optionsDisableScript :: Bool

optionsTargetAddresses :: Maybe TargetAddresses

optionsParser :: Parser Options

parseOptions :: IO Options

utxoDbPath :: Options -> Maybe FilePath

datumDbPath :: Options -> Maybe FilePath

scriptTxDbPath :: Options -> Maybe FilePath