Marconi.Index.Utxo

Documentation

addressFilteredRows :: AddressAny -> [UtxoEvent] -> [UtxoRow]

only store rows in the address list.

newtype Depth

Constructors

Depth Int

eventAtAddress :: AddressAny -> UtxoEvent -> [UtxoEvent]

getUtxos :: IsCardanoEra era => Maybe TargetAddresses -> Tx era -> [Utxo]

getUtxoEvents

Arguments

:: IsCardanoEra era

=> Maybe TargetAddresses
target addresses to filter for

-> SlotNo

-> BlockNo

-> [Tx era]

-> Maybe UtxoEvent
UtxoEvents are stored in storage after conversion to UtxoRow

open

Arguments

:: FilePath
sqlite file path

-> Depth

-> IO UtxoIndex

query

Arguments

:: UtxoIndex
in-memory indexer

-> AddressAny
Address to filter for

-> [UtxoEvent]
volatile events that may be rollbacked

-> IO Result
search results

Query the data stored in the indexer Quries SQL + buffered data, where buffered data is the data that will be batched to SQL

queryPlusVolatile

Arguments

:: UtxoIndex
in-memory indexer

-> AddressAny
Address to filter for

-> IO Result
search results

Query the data stored in the indexer as a whole from: + volatile : in-memory, datat that may rollback + diskStore : on-disk + buffered : in-memeoy, data that will flush to storage

type Result = Maybe [UtxoRow]

toRows :: UtxoEvent -> [UtxoRow]

convert utoEvents to utxoRows

toAddr :: AddressInEra CurrentEra -> AddressAny

utxoAddress :: Lens' Utxo AddressAny

utxoEventSlotNo :: Lens' UtxoEvent SlotNo

utxoEventUtxos :: Lens' UtxoEvent [Utxo]

utxoRowUtxo :: Lens' UtxoRow Utxo

data Utxo

Constructors

Utxo

Fields

_utxoAddress :: !AddressAny

_utxoTxId :: !TxId

_utxoTxIx :: !TxIx

_utxoDatum :: Maybe ScriptData

_utxoDatumHash :: Maybe (Hash ScriptData)

_utxoValue :: Value

_utxoInlineScript :: Maybe ByteString

_utxoInlineScriptHash :: Maybe ScriptHash

data UtxoEvent

Constructors

UtxoEvent

Fields

_utxoEventUtxos :: [Utxo]

_utxoEventInputs :: !(Set TxIn)

_utxoEventSlotNo :: !SlotNo

_utxoEventBlockNo :: !BlockNo

data UtxoRow

Constructors

UtxoRow

Fields

_utxoRowUtxo :: Utxo

_utxoRowSlotNo :: !SlotNo

_utxoRowBlockNo :: !BlockNo

type UtxoIndex = SqliteIndex UtxoEvent Notification AddressAny Result

newtype BlockNo

The 0-based index of the block in the blockchain. BlockNo is <= SlotNo and is only equal at slot N if there is a block for every slot where N <= SlotNo.

Constructors

BlockNo

Fields

unBlockNo :: Word64

newtype SlotNo

The 0-based index for the Ourboros time slot.

Constructors

SlotNo

Fields

unSlotNo :: Word64

Orphan instances

Serialise ScriptData

Instance details

Methods

encode :: ScriptData -> Encoding

decode :: Decoder s ScriptData

encodeList :: [ScriptData] -> Encoding

decodeList :: Decoder s [ScriptData]

FromRow TxIn

Instance details

Methods

fromRow :: RowParser TxIn

FromField Value

Instance details

Methods

fromField :: FieldParser Value

FromField AddressAny

Instance details

Methods

fromField :: FieldParser AddressAny

FromField ScriptHash

Instance details

Methods

fromField :: FieldParser ScriptHash

FromField ScriptData

Instance details

Methods

fromField :: FieldParser ScriptData

FromField TxId

Instance details

Methods

fromField :: FieldParser TxId

FromField TxIx

Instance details

Methods

fromField :: FieldParser TxIx

FromField SlotNo

Instance details

Methods

fromField :: FieldParser SlotNo

FromField BlockNo

Instance details

Methods

fromField :: FieldParser BlockNo

ToRow TxIn

Instance details

Methods

toRow :: TxIn -> [SQLData]

ToField Value

Instance details

Methods

toField :: Value -> SQLData

ToField AddressAny

Instance details

Methods

toField :: AddressAny -> SQLData

ToField ScriptHash

Instance details

Methods

toField :: ScriptHash -> SQLData

ToField ScriptData

Instance details

Methods

toField :: ScriptData -> SQLData

ToField TxId

Instance details

Methods

toField :: TxId -> SQLData

ToField TxIx

Instance details

Methods

toField :: TxIx -> SQLData

ToField SlotNo

Instance details

Methods

toField :: SlotNo -> SQLData

ToField BlockNo

Instance details

Methods

toField :: BlockNo -> SQLData

FromField (Hash ScriptData)

Instance details

Methods

fromField :: FieldParser (Hash ScriptData)

ToField (Hash ScriptData)

Instance details

Methods

toField :: Hash ScriptData -> SQLData