Marconi.Indexers

Documentation

getDatums :: BlockInMode CardanoMode -> [(SlotNo, (Hash ScriptData, ScriptData))]

scriptDataFromCardanoTxBody :: TxBody era -> Map (Hash ScriptData) ScriptData

data Coordinator

The way we synchronise channel consumption is by waiting on a QSemN for each of the spawn indexers to finish processing the current event.

The channel is used to transmit the next event to the listening indexers. Note that even if the channel is unbound it will actually only ever hold one event because it will be blocked until the processing of the event finishes on all indexers.

The indexer count is where we save the number of running indexers so we know for how many we are waiting.

Constructors

Coordinator

Fields

_channel :: TChan (ChainSyncEvent (BlockInMode CardanoMode))

_barrier :: QSemN

_indexerCount :: Int

initialCoordinator :: Int -> IO Coordinator

type Worker = Coordinator -> FilePath -> IO [StorablePoint ScriptTxHandle]

datumWorker :: Worker

isInTargetTxOut

Arguments

:: TargetAddresses
non empty list of target address

-> TxOut CtxTx era
a cardano transaction out that contains an address

-> Bool

does the transaction contain a targetAddress

utxoWorker

Arguments

:: (UtxoIndex -> IO UtxoIndex)
CPS function used in the queryApi thread, needs to be non-blocking

-> Maybe TargetAddresses
Target addresses to filter for

-> Worker

scriptTxWorker_ :: (StorableEvent ScriptTxHandle -> IO [()]) -> Depth -> Coordinator -> TChan (ChainSyncEvent (BlockInMode CardanoMode)) -> FilePath -> IO (IO (), MVar ScriptTxIndexer)

scriptTxWorker :: (StorableEvent ScriptTxHandle -> IO [()]) -> Worker

newtype UtxoQueryTMVar

Constructors

UtxoQueryTMVar

Fields

unUtxoIndex :: TMVar UtxoIndex

for query thread to access in-memory utxos

filterIndexers :: Maybe FilePath -> Maybe FilePath -> Maybe FilePath -> Maybe TargetAddresses -> [(Worker, FilePath)]

startIndexers :: [(Worker, FilePath)] -> IO ([ChainPoint], Coordinator)

mkIndexerStream :: Coordinator -> Stream (Of (ChainSyncEvent (BlockInMode CardanoMode))) IO r -> IO ()