Marconi.Types

Description

This module provides several type aliases and utility functions to deal with them.

Addresses alias used to query marconi

type TargetAddresses = NonEmpty (Address ShelleyAddr)

Typre represents non empty list of Bech32 Shelley compatable addresses

Aliases for the current Cardano era

type CurrentEra = BabbageEra

An alias for the current era, to ease the transition from one era to the next one

pattern AsCurrentEra :: AsType CurrentEra

pattern CurrentEra :: CardanoEra CurrentEra

type TxOut = TxOut CtxTx CurrentEra

A Cardano TxOut of the current Era

Aliases to ease concept mapping between plutus types and cardano types

type TxOutRef = TxIn

A reference to a transaction output. This is a pair of a transaction reference, and an index indicating which of the outputs of that transaction we are referring to.

txOutRef :: TxId -> TxIx -> TxIn